// src/api/generated/model/databaseItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { Time } from './time';

export interface DatabaseItem {
  created_at?: Time;
  created_by?: string;
  description?: string;
  document_id?: string;
  icon?: string;
  id?: string;
  name?: string;
  row_count?: number;
  type?: string;
  updated_at?: Time;
}