// src/api/generated/model/drawingItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { Time } from './time';

export interface DrawingItem {
  created_at?: Time;
  created_by?: string;
  document_id?: string;
  icon?: string;
  id?: string;
  name?: string;
  thumbnail?: string;
  updated_at?: Time;
}